scraper/requirements.txt
scrapyd==1.3.0
Scrapy==2.7.1
scrapy-splash==0.8.0
python-dotenv~=0.21.1
psycopg2-binary==2.9.5
redis==4.4.2

